Redirects successfully after login
