+++ package.json
@@ -21,6 +21,7 @@
     "@react-navigation/native": "^5.8.10",
     "@react-navigation/stack": "^5.12.8",
     "@reduxjs/toolkit": "^1.5.1",
+    "audit": "^0.0.6",
     "axios": "^0.24.0",
     "expo": "^47.0.0",
     "expo-application": "~5.0.1",
@@ -36,12 +37,13 @@
     "expo-web-browser": "^12.0.0",
     "ffmpeg": "^0.0.4",
     "firebase": "^9.16.0",
+    "fix": "^0.0.6",
     "fluent-ffmpeg": "^2.1.2",
     "global": "^4.4.0",
     "i": "^0.3.7",
     "jest": "^26.6.3",
     "moment": "^2.29.2",
-    "npm": "^9.4.0",
+    "npm": "^9.4.2",
     "prop-types": "^15.7.2",
     "react": "18.1.0",
     "react-dom": "18.1.0",
